+++ iscedxreports/admin.py
@@ -1,4 +1,1 @@
-from django.contrib import admin
-from .models import InterSystemsUserProfile
-
-admin.site.register(InterSystemsUserProfile)
+"""Django admin code for iscedxreports Django app."""